+++ include/Events/LevelChangeEventData.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <string>
+#include "EventDataInterface.hpp"
+
+using std::string;
+
+class LevelChangeEventData : public EventDataInterface{
+public:
+  int level;
+  float timeStamp;
+
+  LevelChangeEventData (int level, float timeStamp);
+};